Extract rounding to util function
